+++ app/controllers/tickets_controller.rb
@@ -98,7 +98,11 @@ class TicketsController < ApplicationController
 
   def new
     unless current_user.blank?
-      signature = { content: '<p></p>' + current_user.signature.to_s }
+      if current_user.prefer_plain_text?
+        signature = { content: "\n#{current_user.signature}" }
+      else
+        signature = { content: "<p></p>#{current_user.signature}" }
+      end
     else
       signature = {}
     end
@@ -192,6 +196,7 @@ class TicketsController < ApplicationController
             :assignee_id,
             :priority,
             :message_id,
+            :content_type,
             attachments_attributes: [
               :file
             ])
@@ -201,6 +206,7 @@ class TicketsController < ApplicationController
             :content,
             :subject,
             :priority,
+            :content_type,
             attachments_attributes: [
               :file
             ])