+++ db/migrate/20150821131046_add_plain_text_preference_to_user.rb
@@ -1,0 +1,5 @@
+class AddPlainTextPreferenceToUser < ActiveRecord::Migration
+  def change
+    add_column :users, :prefer_plain_text, :boolean, default: false, null: false
+  end
+end