+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Opt-in other users in conversations.
 - Filter users by type and/or email.
 - Possible to inline-edit the label name.
+- Per user configuration option for plain text replies.
 
 ### Changed
 - The outgoing email address of replies will now be the same as the original incoming email addresses when it was configured correctly as an outgoing email address. The agent can choose a different address when replying.